+++ src/utils/event-bus.js
@@ -21,7 +21,10 @@ export const events = {
   USER_INPUT_FOCUS_CHANGED: 'user_input_focus_changed',
   SCROLL_CHAT_DOWN: 'scroll_chat_down',
   SET_ENGINE_URL: "set-engine-url",
-  SET_LOCALE: 'set_locale'
+  SET_LOCALE: 'set_locale',
+  SHOW_CALLOUT: 'show_callout',
+  HIDE_CALLOUT: 'hide_calllout',
+  ZOOM_IMAGE: 'zoom_image'
 };
 
 export const EventBus = new Vue();