+++ src/utils/message-list-cache.js
@@ -1,12 +1,13 @@
+import { store } from '../store/store.js';
 class MessageListCache {
   constructor() {
-    this.enabled = !!window.sessionStorage;
+    this.enabled = !!store.getters.storage;
     this.cacheKey = 'teneo-web-chat-messages';
   }
 
   get() {
     if (this.enabled) {
-      const cached = window.sessionStorage.getItem(this.cacheKey);
+      const cached = store.getters.storage.getItem(this.cacheKey);
 
       if (!cached) {
         return [];
@@ -18,7 +19,7 @@ class MessageListCache {
 
   update(value) {
     if (this.enabled) {
-      window.sessionStorage.setItem(this.cacheKey, JSON.stringify(value));
+      store.getters.storage.setItem(this.cacheKey, JSON.stringify(value));
     }
   }
 }