+++ src/utils/validate-url.js
@@ -1,19 +1,21 @@
 export default function isValidURL(str) {
-    if (isReglarURL(str) || isDataURL(str)) {
+    if (isRegularURL(str) || isDataURL(str)) {
         return true
     } else {
         return false
     }
 }
 
-function isReglarURL(str) {
-    var pattern = new RegExp('^(https?:\\/\\/)?'+ // protocol
-    '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.?)+[a-z]{2,}|'+ // domain name
-    '((\\d{1,3}\\.){3}\\d{1,3}))'+ // ip (v4) address
-    '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*'+ //port
-    '(\\?[;&amp;a-z\\d%_.~+=-]*)?'+ // query string
-    '(\\#[-a-z\\d_]*)?$','i');
-    return pattern.test(str);
+function isRegularURL(string) {
+    let url;
+  
+    try {
+      url = new URL(string);
+    } catch (_) {
+      return false;  
+    }
+  
+    return url.protocol === "http:" || url.protocol === "https:";
 }
 
 // Detecting data URLs